src/requirements.txt
gino==1.0
gino[starlette]==1.0
fastapi==0.63.0
uvicorn==0.13.4
alembic==1.6.5
pydantic
pydantic[email]
pyjwt
passlib[bcrypt]
python-multipart
gunicorn==20.1.0
sqlalchemy==1.3.15
psycopg2-binary==2.8.6

# dev
pytest==6.2.3
pytest-asyncio
httpx
requests==2.25.1
faker==8.7.0
